+++ backend/src/main/java/erp_backend/entity/CreditCollection.java
@@ -1,5 +1,6 @@
 package erp_backend.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import jakarta.persistence.Column;
@@ -24,7 +25,7 @@ public class CreditCollection {
     private Invoice invoice;
 
     @Column(name = "AmountCollected", nullable = false)
-    private Double amountCollected;
+    private BigDecimal amountCollected;
 
     @Column(name = "CollectionDate", nullable = false)
     private LocalDate collectionDate;
@@ -33,8 +34,8 @@ public class CreditCollection {
     public void setCollectionId(Integer collectionId) { this.collectionId = collectionId; }
     public Invoice getInvoice() { return invoice; }
     public void setInvoice(Invoice invoice) { this.invoice = invoice; }
-    public Double getAmountCollected() { return amountCollected; }
-    public void setAmountCollected(Double amountCollected) { this.amountCollected = amountCollected; }
+    public BigDecimal getAmountCollected() { return amountCollected; }
+    public void setAmountCollected(BigDecimal amountCollected) { this.amountCollected = amountCollected; }
     public LocalDate getCollectionDate() { return collectionDate; }
     public void setCollectionDate(LocalDate collectionDate) { this.collectionDate = collectionDate; }
 }
